# Conflicts:
#	Application-Module/src/main/kotlin/com/clip/application/store/port/out/ZoneManagementPort.kt
#	Application-Module/src/main/kotlin/com/clip/application/store/service/StoreQueryService.kt
#	Domain-Module/src/main/kotlin/com/clip/domain/store/entity/Zone.kt
#	Infrastructure-Module/Persistence/src/main/kotlin/com/clip/persistence/jpa/store/ZoneMapper.kt
#	Infrastructure-Module/Persistence/src/main/kotlin/com/clip/persistence/jpa/store/adapter/ZoneManagementJpaAdapter.kt

+++ Infrastructure-Module/Persistence/src/main/kotlin/com/clip/persistence/jpa/store/entity/StoreMenuEntity.kt
@@ -10,14 +10,14 @@ class StoreMenuEntity(
     id : String,
     storeId: String,
     name: String,
-    price: Double
+    price: Long,
 ) : AggregateRoot<StoreMenuEntity>(id) {
 
     @Column(nullable = false, length = 255)
     var name: String = name
 
     @Column(nullable = false)
-    var price: Double = price
+    var price: Long = price
 
     @Column(nullable = false)
     val storeId: String = storeId
